feat: hide logic for inputs in hooks
